+++ app/src/test/java/com/example/sudoku/domain/usecase/RemoveDigitsUseCaseTest.kt
@@ -1,7 +1,5 @@
 package com.example.sudoku.domain.usecase
 
-import com.example.sudoku.domain.usecase.GenerateGridUseCase
-import com.example.sudoku.domain.usecase.RemoveDigitsUseCase
 import org.junit.Test
 import kotlin.test.assertEquals
 
@@ -19,7 +17,7 @@ class RemoveDigitsUseCaseTest {
         val removedDigits = grid.sumOf { row ->
             var counter = 0
             row.forEach { digit ->
-                if (digit.value == 0.toShort())
+                if (digit.value == 0)
                     counter += 1
             }
             counter
